Refactor YandexMerchant extension with controllers
